Add telemetry for prompt after create
